fix: address code review findings on --catchup mode
- disk-layout: latestOnDisk now constrains discovery to canonical depth
  (root/YYYY/MM/DD/H.json.gz). Stray .json.gz files at other depths would
  otherwise yield malformed hour IDs and silently poison the anchor —
  a NaN seed in the reduce would persist and the catchup walk would
  silently no-op.
- disk-layout: export hourIdToMs / msToHourId as single source of truth
  for hour-arithmetic. fetcher.js now imports them instead of carrying
  duplicate private copies.
- fetcher: cold-start now throws instead of calling process.exit(1)
  directly, so the main().catch fatal-error path handles it consistently
  and pino's transport worker gets a chance to flush before exit.

Refs PR #195 review (orchestrator security/correctness pass).
